requirements.txt
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.9.2
Fabric3==1.14.post1
gitdb==4.0.5
paramiko==2.7.1
pycparser==2.20
PyNaCl==1.3.0
six==1.14.0
smmap==3.0.4
